Refactor: Decouple agent execution and tools
