+++ src/coding_assistant/tools/__init__.py
@@ -1,0 +1,1 @@
+# Tools package - contains all tool-related functionality
